+++ cedar-policy-validator/src/schema_file_format.rs
@@ -15,10 +15,7 @@
  */
 
 use cedar_policy_core::entities::JSONValue;
-use serde::{
-    de::{MapAccess, Visitor},
-    Deserialize, Serialize,
-};
+use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 use smol_str::SmolStr;
 use std::collections::{BTreeMap, HashMap};
@@ -184,7 +181,7 @@ impl std::fmt::Display for ActionEntityUID {
 
 /// A restricted version of the `Type` enum containing only the types which are
 /// exposed to users.
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
 // This enum is `untagged` with these variants as a workaround to a serde
 // limitation. It is not possible to have the known variants on one enum, and
 // then, have catch-all variant for any unrecognized tag in the same enum that
@@ -198,213 +195,15 @@ pub enum SchemaType {
     },
 }
 
-impl<'de> Deserialize<'de> for SchemaType {
-    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        deserializer.deserialize_any(SchemaTypeVisitor)
-    }
-}
-
-/// The fields for a `SchemaTypes`. Used for implementing deserialization.
-#[derive(Hash, Eq, PartialEq, Deserialize)]
-#[serde(field_identifier, rename_all = "camelCase")]
-enum TypeFields {
-    Type,
-    Element,
-    Attributes,
-    AdditionalAttributes,
-    Name,
-}
-
-// This macro is used to avoid duplicating the fields names when calling
-// `serde::de::Error::unknown_field`. It wants an `&'static [&'static str]`, and
-// AFAIK, the elements of the static slice must be literals.
-macro_rules! type_field_name {
-    (Type) => {
-        "type"
-    };
-    (Element) => {
-        "element"
-    };
-    (Attributes) => {
-        "attributes"
-    };
-    (AdditionalAttributes) => {
-        "additionalAttributes"
-    };
-    (Name) => {
-        "name"
-    };
-}
-
-impl TypeFields {
-    fn as_str(&self) -> &'static str {
-        match self {
-            TypeFields::Type => type_field_name!(Type),
-            TypeFields::Element => type_field_name!(Element),
-            TypeFields::Attributes => type_field_name!(Attributes),
-            TypeFields::AdditionalAttributes => type_field_name!(AdditionalAttributes),
-            TypeFields::Name => type_field_name!(Name),
-        }
-    }
-}
-
-/// Used during deserialization to deserialize the attributes type map while
-/// reporting an error if there are any duplicate keys in the map. I could not
-/// find a way to do the `serde_with` conversion inline without introducing this
-/// struct.
-#[derive(Deserialize)]
-struct AttributesTypeMap(
-    #[serde(with = "serde_with::rust::maps_duplicate_key_is_error")]
-    BTreeMap<SmolStr, TypeOfAttribute>,
-);
-
-struct SchemaTypeVisitor;
-
-impl<'de> Visitor<'de> for SchemaTypeVisitor {
-    type Value = SchemaType;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-        formatter.write_str("builtin type or reference to type defined in commonTypes")
-    }
-
-    fn visit_map<M>(self, mut map: M) -> std::result::Result<Self::Value, M::Error>
-    where
-        M: MapAccess<'de>,
-    {
-        use TypeFields::*;
-
-        // We keep field values wrapped in a `Result` initially so that we do
-        // not report errors due the contents of a field when the field is not
-        // expected for a particular type variant. We instead report that the
-        // field so not exist at all, so that the schema author can delete the
-        // field without wasting time fixing errors in the value.
-        let mut type_name: Option<std::result::Result<SmolStr, M::Error>> = None;
-        let mut element: Option<std::result::Result<SchemaType, M::Error>> = None;
-        let mut attributes: Option<std::result::Result<AttributesTypeMap, M::Error>> = None;
-        let mut additional_attributes: Option<std::result::Result<bool, M::Error>> = None;
-        let mut name: Option<std::result::Result<SmolStr, M::Error>> = None;
-
-        // Gather all the fields in the object. Any fields that are not one of
-        // the possible fields for some schema type will have been reported by
-        // serde already.
-        while let Some(key) = map.next_key()? {
-            match key {
-                Type => {
-                    if type_name.is_some() {
-                        return Err(serde::de::Error::duplicate_field(Type.as_str()));
-                    }
-                    type_name = Some(map.next_value());
-                }
-                Element => {
-                    if element.is_some() {
-                        return Err(serde::de::Error::duplicate_field(Element.as_str()));
-                    }
-                    element = Some(map.next_value());
-                }
-                Attributes => {
-                    if attributes.is_some() {
-                        return Err(serde::de::Error::duplicate_field(Attributes.as_str()));
-                    }
-                    attributes = Some(map.next_value());
-                }
-                AdditionalAttributes => {
-                    if additional_attributes.is_some() {
-                        return Err(serde::de::Error::duplicate_field(
-                            AdditionalAttributes.as_str(),
-                        ));
-                    }
-                    additional_attributes = Some(map.next_value());
-                }
-                Name => {
-                    if name.is_some() {
-                        return Err(serde::de::Error::duplicate_field(Name.as_str()));
-                    }
-                    name = Some(map.next_value());
-                }
-            }
-        }
-
-        Self::build_schema_type::<M>(type_name, element, attributes, additional_attributes, name)
-    }
-}
-
-impl SchemaTypeVisitor {
-    /// Construct a schema type given the name of the type and its fields.
-    /// Fields which were not present are `None`. It is an error for a field
-    /// which is not used for a particular type to be `Some` when building that
-    /// type.
-    fn build_schema_type<'de, M>(
-        type_name: Option<std::result::Result<SmolStr, M::Error>>,
-        element: Option<std::result::Result<SchemaType, M::Error>>,
-        attributes: Option<std::result::Result<AttributesTypeMap, M::Error>>,
-        additional_attributes: Option<std::result::Result<bool, M::Error>>,
-        name: Option<std::result::Result<SmolStr, M::Error>>,
-    ) -> std::result::Result<SchemaType, M::Error>
-    where
-        M: MapAccess<'de>,
-    {
-        use TypeFields::*;
-
-        match type_name.transpose()?.as_ref().map(|s| s.as_str()) {
-            Some("String") => Ok(SchemaType::Type(SchemaTypeVariant::String)),
-            Some("Long") => Ok(SchemaType::Type(SchemaTypeVariant::Long)),
-            Some("Boolean") => Ok(SchemaType::Type(SchemaTypeVariant::Boolean)),
-            Some("Set") => {
-                if let Some(element) = element {
-                    Ok(SchemaType::Type(SchemaTypeVariant::Set {
-                        element: Box::new(element?),
-                    }))
-                } else {
-                    Err(serde::de::Error::missing_field(Element.as_str()))
-                }
-            }
-            Some("Record") => {
-                if let Some(attributes) = attributes {
-                    let additional_attributes =
-                        additional_attributes.unwrap_or(Ok(additional_attributes_default()));
-                    Ok(SchemaType::Type(SchemaTypeVariant::Record {
-                        attributes: attributes?.0,
-                        additional_attributes: additional_attributes?,
-                    }))
-                } else {
-                    Err(serde::de::Error::missing_field(Attributes.as_str()))
-                }
-            }
-            Some("Entity") => {
-                if let Some(name) = name {
-                    Ok(SchemaType::Type(SchemaTypeVariant::Entity { name: name? }))
-                } else {
-                    Err(serde::de::Error::missing_field(Name.as_str()))
-                }
-            }
-            Some("Extension") => {
-                if let Some(name) = name {
-                    Ok(SchemaType::Type(SchemaTypeVariant::Extension {
-                        name: name?,
-                    }))
-                } else {
-                    Err(serde::de::Error::missing_field(Name.as_str()))
-                }
-            }
-            Some(type_name) => Ok(SchemaType::TypeDef {
-                type_name: type_name.into(),
-            }),
-            None => Err(serde::de::Error::missing_field(Type.as_str())),
-        }
-    }
-}
-
 impl From<SchemaTypeVariant> for SchemaType {
     fn from(variant: SchemaTypeVariant) -> Self {
         Self::Type(variant)
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
 #[serde(tag = "type")]
+#[serde(deny_unknown_fields)]
 pub enum SchemaTypeVariant {
     String,
     Long,
@@ -413,8 +212,10 @@ pub enum SchemaTypeVariant {
         element: Box<SchemaType>,
     },
     Record {
+        #[serde(with = "serde_with::rust::maps_duplicate_key_is_error")]
         attributes: BTreeMap<SmolStr, TypeOfAttribute>,
         #[serde(rename = "additionalAttributes")]
+        #[serde(default = "additional_attributes_default")]
         additional_attributes: bool,
     },
     Entity {
@@ -550,6 +351,8 @@ fn record_attribute_required_default() -> bool {
 
 #[cfg(test)]
 mod test {
+    use crate::ValidatorSchema;
+
     use super::*;
 
     #[test]
@@ -697,7 +500,6 @@ mod test {
     }
 
     #[test]
-    #[should_panic(expected = "unknown field `requiredddddd`")]
     fn test_schema_file_with_misspelled_required() {
         let src = serde_json::json!(
         {
@@ -722,7 +524,6 @@ mod test {
     }
 
     #[test]
-    #[should_panic(expected = "unknown field `nameeeeee`")]
     fn test_schema_file_with_misspelled_field() {
         let src = serde_json::json!(
         {
@@ -746,7 +547,6 @@ mod test {
     }
 
     #[test]
-    #[should_panic(expected = "unknown field `extra`")]
     fn test_schema_file_with_extra_field() {
         let src = serde_json::json!(
         {
@@ -771,7 +571,6 @@ mod test {
     }
 
     #[test]
-    #[should_panic(expected = "unknown field `memberOfTypes`")]
     fn test_schema_file_with_misplaced_field() {
         let src = serde_json::json!(
         {
@@ -796,7 +595,10 @@ mod test {
     }
 
     #[test]
-    #[should_panic(expected = "missing field `name`")]
+    // This schema was rejected prior to the reverted schema parsing changes.
+    // The error message was improved by the changes, so the current message is
+    // not very helpful.
+    #[should_panic(expected = "UndeclaredCommonTypes({\"Entity\"})")]
     fn schema_file_with_missing_field() {
         let src = serde_json::json!(
         {
@@ -816,10 +618,14 @@ mod test {
         });
         let schema: NamespaceDefinition = serde_json::from_value(src).unwrap();
         println!("{:#?}", schema);
+        TryInto::<ValidatorSchema>::try_into(schema).unwrap();
     }
 
     #[test]
-    #[should_panic(expected = "missing field `type`")]
+    // This schema was rejected prior to the reverted schema parsing changes.
+    // The error message was improved by the changes, so the current message is
+    // not very helpful.
+    #[should_panic(expected = "data did not match any variant of untagged enum SchemaType")]
     fn schema_file_with_missing_type() {
         let src = serde_json::json!(
         {
@@ -963,7 +769,6 @@ mod test_duplicates_error {
     }
 
     #[test]
-    #[should_panic(expected = "invalid entry: found duplicate key")]
     fn record_type() {
         let src = r#"{
             "Foo": {